更新角色、资源权限； 角色形象列表

+++ omgcms/src/main/java/com/omgcms/repository/RoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import com.omgcms.model.core.Role;
 
-public interface RoleRepository extends JpaSpecificationExecutor<Role>, JpaRepository<Role, Integer>{
+public interface RoleRepository extends JpaSpecificationExecutor<Role>, JpaRepository<Role, Long>{
 	
 	
 }
